Fix Denial of Service bug equivalent to CVE-2025-27144
This change fixes a denial of service bug equivalent to CVE-2025-27144.

The issue is that a malicious attacker could present JWTs and JWEs with
excessive dot separators. The original implementation is subject
to excessive memory use due to the use of the 'strings.Split' function
which will continue to split compact-encoded JWTs and JWEs beyond the
expected number. It is expected that the sevrity of this issue is low.

Secondy, this PR also adds a timeout to the internal JWKS fetcher.
Previously, the original solution used the default getter from the Go
standard http package which has an unlimited timeout. A malicious JWKS
server could keep a connection alive indefinately and never return
stalling a Go routine forever. This could lead to resource exhaustion
and DoS conditions. Given a JWKS URL is most often configured as
trusted, this is alos a low severity issue.

Finally, this PR fixes a broken unit test swapping SHA256/SHA1 in a
known answer test.

Signed-off-by: Nick Smith <nickrmc83@gmail.com>